refactor: 파일 구조 설정
