+++ mejai/mejai/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
