+++ mejai/mejai/Resource/Assets.xcassets/Contents.json
